Separate out responsibility endpoints
