+++ src/app/src/main/kotlin/org/vaccineimpact/api/app/controllers/ModellingGroupController.kt
@@ -43,9 +43,6 @@ open class ModellingGroupController(context: ControllerContext)
         return listOf(
                 oneRepoEndpoint("/", this::getModellingGroups, repos, repo).secured(setOf("*/modelling-groups.read")),
                 oneRepoEndpoint("/:group-id/", this::getModellingGroup, repos, repo).secured(setOf("*/modelling-groups.read", "*/models.read")),
-                oneRepoEndpoint("$responsibilitiesURL/", this::getResponsibilities, repos, repo).secured(responsibilityPermissions),
-                oneRepoEndpoint("$touchstonesURL/", this::getTouchstones, repos, repo).secured(touchtonePermissions),
-                oneRepoEndpoint("$scenarioURL/", this::getResponsibility, repos, repo).secured(responsibilityPermissions),
                 oneRepoEndpoint("/:group-id/actions/associate_member/", this::modifyMembership, repos, { it.user }, method = HttpMethod.post).secured()
         )
     }
@@ -60,34 +57,6 @@ open class ModellingGroupController(context: ControllerContext)
     {
         val groupId = groupId(context)
         return repo.getModellingGroupDetails(groupId)
-    }
-
-    fun getResponsibilities(context: ActionContext, repo: ModellingGroupRepository): Responsibilities
-    {
-        val groupId = groupId(context)
-        val touchstoneId = context.params(":touchstone-id")
-        val filterParameters = ScenarioFilterParameters.fromContext(context)
-
-        val data = repo.getResponsibilities(groupId, touchstoneId, filterParameters)
-        context.checkIsAllowedToSeeTouchstone(touchstoneId, data.touchstoneStatus)
-        return data.responsibilities
-    }
-
-    fun getTouchstones(context: ActionContext, repo: ModellingGroupRepository): List<Touchstone>
-    {
-        val groupId = groupId(context)
-
-        var touchstones = repo.getTouchstonesByGroupId(groupId)
-        touchstones = touchstones.filter { context.isAllowedToSeeTouchstone(it.status) }
-        return touchstones
-    }
-
-    fun getResponsibility(context: ActionContext, repo: ModellingGroupRepository): ResponsibilityAndTouchstone
-    {
-        val path = ResponsibilityPath(context)
-        val data = repo.getResponsibility(path.groupId, path.touchstoneId, path.scenarioId)
-        context.checkIsAllowedToSeeTouchstone(path.touchstoneId, data.touchstone.status)
-        return data
     }
 
     fun modifyMembership(context: ActionContext, repo: UserRepository): String
