+++ src/app/src/test/kotlin/org/vaccineimpact/api/tests/controllers/ModellingGroupControllersTests.kt
@@ -27,73 +27,6 @@ class ModellingGroupControllersTests : ControllerTests<ModellingGroupController>
             = ModellingGroupController(controllerContext)
 
     @Test
-    fun `getResponsibilities gets parameters from URL`()
-    {
-        val data = ResponsibilitiesAndTouchstoneStatus(
-                Responsibilities("tId", "", null, emptyList()),
-                TouchstoneStatus.FINISHED
-        )
-        val repo = mock<ModellingGroupRepository> {
-            on { getResponsibilities(any(), any(), any()) } doReturn data
-        }
-        val context = mock<ActionContext> {
-            on { it.params(":group-id") } doReturn "gId"
-            on { it.params(":touchstone-id") } doReturn "tId"
-            on { hasPermission(any()) } doReturn true
-        }
-
-        val controller = ModellingGroupController(mockControllerContext())
-        controller.getResponsibilities(context, repo)
-
-        verify(repo).getResponsibilities(eq("gId"), eq("tId"), any())
-    }
-
-    @Test
-    fun `getResponsibilities returns error if user does not have permission to see in-preparation touchstone`()
-    {
-        val data = ResponsibilitiesAndTouchstoneStatus(
-                Responsibilities("tId", "", null, emptyList()),
-                TouchstoneStatus.IN_PREPARATION
-        )
-        val repo = mock<ModellingGroupRepository> {
-            on { getResponsibilities(any(), any(), any()) } doReturn data
-        }
-        val context = mock<ActionContext> {
-            on { it.params(":group-id") } doReturn "gId"
-            on { it.params(":touchstone-id") } doReturn "tId"
-            on { hasPermission(any()) } doReturn false
-        }
-
-        val controller = ModellingGroupController(mockControllerContext())
-        assertThatThrownBy {
-            controller.getResponsibilities(context, repo)
-        }.hasMessageContaining("Unknown touchstone")
-    }
-
-    @Test
-    fun `getResponsibility gets parameters from URL`()
-    {
-        val repo = makeRepoMockingGetResponsibility(TouchstoneStatus.OPEN)
-        val context = mockContextForSpecificResponsibility(true)
-        val controller = ModellingGroupController(mockControllerContext())
-        controller.getResponsibility(context, repo)
-
-        verify(repo).getResponsibility(eq("gId"), eq("tId"), eq("sId"))
-    }
-
-    @Test
-    fun `getResponsibility returns error if user does not have permission to see in-preparation touchstone`()
-    {
-        val repo = makeRepoMockingGetResponsibility(TouchstoneStatus.IN_PREPARATION)
-        val context = mockContextForSpecificResponsibility(false)
-        val controller = ModellingGroupController(mockControllerContext())
-        assertThatThrownBy { controller.getResponsibility(context, repo) }
-                .hasMessageContaining("Unknown touchstone")
-    }
-
-
-
-    @Test
     fun `modifyMembership returns error if user does not have permission to manage members`()
     {
         val context = mock<ActionContext> {
@@ -132,72 +65,4 @@ class ModellingGroupControllersTests : ControllerTests<ModellingGroupController>
         val controller = ModellingGroupController(mockControllerContext())
         controller.modifyMembership(context, mock<UserRepository>())
     }
-
-    @Test
-    fun `returns in preparation touchstones if user has permission to read prepared touchstones`()
-    {
-        val groupId = "test-group"
-        val repo = mock<ModellingGroupRepository> {
-            on { getTouchstonesByGroupId(groupId) } doReturn mockTouchstones
-        }
-
-        val controller = makeController(mockControllerContext())
-
-        val context = mock<ActionContext> {
-            on { params(":group-id") } doReturn groupId
-            on { hasPermission(ReifiedPermission.parse("*/touchstones.prepare")) } doReturn true
-        }
-        val data = controller.getTouchstones(context, repo)
-        assertThat(data.count()).isEqualTo(2)
-    }
-
-    @Test
-    fun `does not return in preparation touchstones if user has no permission to read prepared touchstones`()
-    {
-        val groupId = "test-group"
-        val repo = mock<ModellingGroupRepository> {
-            on { getTouchstonesByGroupId(groupId) } doReturn mockTouchstones
-        }
-
-        val controller = makeController(mockControllerContext())
-
-        val context = mock<ActionContext> {
-            on { params(":group-id") } doReturn groupId
-            on { hasPermission(ReifiedPermission.parse("*/touchstones.prepare")) } doReturn false
-        }
-        val data = controller.getTouchstones(context, repo)
-        assertThat(data.count()).isEqualTo(1)
-    }
-
-    private val mockTouchstones = listOf(
-            Touchstone("touchstone-1", "touchstone", 1, "Description", TouchstoneStatus.OPEN),
-            Touchstone("touchstone-bad", "touchstone", 1, "not open", TouchstoneStatus.IN_PREPARATION)
-    )
-
-    private fun mockContextForSpecificResponsibility(hasPermissions: Boolean): ActionContext
-    {
-        val context = mock<ActionContext> {
-            on { it.params(":group-id") } doReturn "gId"
-            on { it.params(":touchstone-id") } doReturn "tId"
-            on { it.params(":scenario-id") } doReturn "sId"
-            on { hasPermission(any()) } doReturn hasPermissions
-        }
-        return context
-    }
-
-    private fun makeRepoMockingGetResponsibility(status: TouchstoneStatus): ModellingGroupRepository
-    {
-        val data = ResponsibilityAndTouchstone(
-                Touchstone("tId", "t", 1, "desc", status),
-                Responsibility(
-                        Scenario("sId", "scDesc", "disease", listOf("t-1")),
-                        ResponsibilityStatus.EMPTY, emptyList(), null
-                )
-        )
-        return mock {
-            on { getResponsibility(any(), any(), any()) } doReturn data
-        }
-    }
-
-
 }
